Move entity to app
